+++ src/main/java/net/more_rpg_classes/client/particle/MoreParticles.java
@@ -28,6 +28,7 @@ public class MoreParticles {
     public static final DefaultParticleType WIND_VACUUM = FabricParticleTypes.simple();
     public static final DefaultParticleType SMALL_GUST = FabricParticleTypes.simple();
     public static final DefaultParticleType GUST = FabricParticleTypes.simple();
+    public static final DefaultParticleType GAS_CLOUD = FabricParticleTypes.simple();
 
 
     public static void register(){
@@ -51,6 +52,7 @@ public class MoreParticles {
         Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "wind_vacuum"), WIND_VACUUM);
         Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "small_gust"), SMALL_GUST);
         Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "gust"), GUST);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "gas_cloud"), GAS_CLOUD);
     }
 
 }