+++ src/main/java/net/more_rpg_classes/effect/FrostedEffect.java
@@ -12,28 +12,12 @@ public class FrostedEffect extends StatusEffect {
         super(statusEffectCategory, color);
     }
 
-
-    @Override
-    public void applyUpdateEffect(LivingEntity pLivingEntity, int pAmplifier) {
-        pLivingEntity.setFrozenTicks(pLivingEntity.getFrozenTicks() + 1);
-        if(pAmplifier == MRPGCMod.effectsConfig.value.frosted_amplifier_frozen_solid_conversion){
-            pLivingEntity.addStatusEffect(new StatusEffectInstance(MRPGCEffects.FROZEN_SOLID,100,0,false,false,true));
-            pLivingEntity.removeStatusEffect(MRPGCEffects.FROSTED);
-        }
-        super.applyUpdateEffect(pLivingEntity, pAmplifier);
-    }
-
-    @Override
-    public boolean canApplyUpdateEffect(int duration, int amplifier) {
-        return true;
-    }
-
     @Override
     public void onApplied(LivingEntity entity, AttributeContainer attributes, int amplifier) {
-        entity.setFrozenTicks(entity.getFrozenTicks() + 40);
-    }
-    @Override
-    public void onRemoved(LivingEntity entity, AttributeContainer attributes, int amplifier) {
-        entity.setFrozenTicks(entity.getFrozenTicks() + 200);
+        entity.setFrozenTicks(entity.getFrozenTicks() + (60 * amplifier));
+        if(amplifier == MRPGCMod.effectsConfig.value.frosted_amplifier_frozen_solid_conversion){
+            entity.addStatusEffect(new StatusEffectInstance(MRPGCEffects.FROZEN_SOLID,100,0,false,false,true));
+            entity.removeStatusEffect(MRPGCEffects.FROSTED);
+        }
     }
 }